Moved seq_to_satn to tools instead of filters
